Introducing useNavigation()
- Use useNavigation() everywhere instead of circular dep on
navigation.ts
- Add types/JwtPayload.ts to remove circular dep
- Add types/VerifyOptions.ts to remove circular dep
