+++ src/index.tsx
@@ -4,17 +4,28 @@ import Toast from "react-native-toast-message";
 import { ColorContextProvider } from "./colorContext";
 import { ContextProvider } from "./context";
 import { Navigation } from "./navigation";
+import {
+    createNavigationContainerRef,
+    NavigationContainer,
+} from "@react-navigation/native";
+import RNBootSplash from "react-native-bootsplash";
+
+const navigationRef = createNavigationContainerRef();
 
 const App = () => {
     return (
-        <SafeAreaProvider>
-            <ColorContextProvider>
-                <ContextProvider>
-                    <Navigation />
-                    <Toast ref={(ref) => Toast.setRef(ref)} />
-                </ContextProvider>
-            </ColorContextProvider>
-        </SafeAreaProvider>
+        <NavigationContainer
+            onReady={() => RNBootSplash.hide({ fade: true })}
+            ref={navigationRef}>
+            <SafeAreaProvider>
+                <ColorContextProvider>
+                    <ContextProvider>
+                        <Navigation />
+                        <Toast ref={(ref) => Toast.setRef(ref)} />
+                    </ContextProvider>
+                </ColorContextProvider>
+            </SafeAreaProvider>
+        </NavigationContainer>
     );
 };
