+++ src/navigation.tsx
@@ -1,9 +1,7 @@
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import { NavigationContainer } from "@react-navigation/native";
 import { createStackNavigator } from "@react-navigation/stack";
 import * as React from "react";
 import { useContext } from "react";
-import RNBootSplash from "react-native-bootsplash";
 import { Icon, IconType } from "./assets/icons/Icon";
 import { ColorContext } from "./colorContext";
 import { Context } from "./context";
@@ -14,25 +12,8 @@ import { ProfileNavigation } from "./screens/Profile/ProfileNavigation";
 import RequestAndProposalHandler from "./screens/RequestAndProposalHandler";
 import { ScannerScreen } from "./screens/ScannerScreen";
 
-export const navigationRef = React.createRef<any>();
-
 const Stack = createStackNavigator();
 const Tab = createBottomTabNavigator();
-
-export function navigate(name: string, params?: any) {
-    if (typeof navigationRef.current === "undefined") {
-        return;
-    }
-    navigationRef.current.navigate(name, params);
-}
-
-export function goBack() {
-    if (typeof navigationRef.current === "undefined") {
-        return;
-    }
-    navigationRef.current.goBack();
-}
-
 interface Route {
     routeId: string;
     title: string;
@@ -96,43 +77,40 @@ function Tabs() {
 export const Navigation = () => {
     const { selectedChain, isTest } = useContext(Context);
     const { colors, toggleDarkMode } = useContext(ColorContext);
+
     // const { colors } = useTheme();
     return (
-        <NavigationContainer
-            onReady={() => RNBootSplash.hide({ fade: true })}
-            ref={navigationRef}>
-            <Stack.Navigator>
-                <Stack.Screen
-                    name="Main"
-                    component={Tabs}
-                    options={{ headerShown: false }}
-                />
-                <Stack.Screen
-                    name="Identity"
-                    component={Identity}
-                    options={{
-                        title: "Identitet",
-                        headerTitleStyle: {
-                            color: colors.onPrimary,
-                        },
-                    }}
-                />
-                <Stack.Screen
-                    name="Bankid"
-                    component={BankId}
-                    options={{ title: "BankID" }}
-                />
-                <Stack.Screen
-                    name="Modal"
-                    component={RequestAndProposalHandler}
-                    options={{ headerShown: false }}
-                />
-                <Stack.Screen
-                    name="Scanner"
-                    component={ScannerScreen}
-                    options={{ title: "Scan QR" }}
-                />
-            </Stack.Navigator>
-        </NavigationContainer>
+        <Stack.Navigator>
+            <Stack.Screen
+                name="Main"
+                component={Tabs}
+                options={{ headerShown: false }}
+            />
+            <Stack.Screen
+                name="Identity"
+                component={Identity}
+                options={{
+                    title: "Identitet",
+                    headerTitleStyle: {
+                        color: colors.onPrimary,
+                    },
+                }}
+            />
+            <Stack.Screen
+                name="Bankid"
+                component={BankId}
+                options={{ title: "BankID" }}
+            />
+            <Stack.Screen
+                name="Modal"
+                component={RequestAndProposalHandler}
+                options={{ headerShown: false }}
+            />
+            <Stack.Screen
+                name="Scanner"
+                component={ScannerScreen}
+                options={{ title: "Scan QR" }}
+            />
+        </Stack.Navigator>
     );
 };
