+++ src/types/JwtPayload.ts
@@ -1,0 +1,10 @@
+export interface JwtPayload {
+    [key: string]: any;
+    iss?: string;
+    sub?: string;
+    aud?: string | string[];
+    exp?: number;
+    nbf?: number;
+    iat?: number;
+    jti?: string;
+}
