+++ src/types/VerifyOptions.ts
@@ -1,0 +1,10 @@
+export interface VerifyOptions {
+    audience: string;
+    complete: boolean;
+    issuer: string | string[];
+    ignoreExpiration: boolean;
+    ignoreNotBefore: boolean;
+    subject: string;
+    decodeCredentials: boolean;
+    requireVerifiablePresentation: boolean;
+}
